DKT-Server/requirements.txt
pandas
pymongo
python-dotenv
fastapi
pydantic
requests
